Inserted MiniUML.Model.MiniUmlPluginLoader.LoadPlugins into  Method Edi.BootStrapper.CreateShell
The method
Edi.BootStrapper.CreateModuleCatalog()

has an error at  PluginModuleCatalog

The type or namespace name 'PluginModuleCatalog' could not be found (are you missing a using directive or an assembly reference?)

but the application does nt work without it at runtime. So, this is needed. Should we ignore the compilation errer or is there something to fix the compilation error?